Move downwardmetrics to /pkg/downwardmetrics
This PR moves downwardmetrics from
/pkg/monitoring to /pkg/downwardmetrics
