+++ cmd/virt-handler/BUILD.bazel
@@ -15,8 +15,8 @@ go_library(
         "//pkg/certificates/bootstrap:go_default_library",
         "//pkg/container-disk:go_default_library",
         "//pkg/controller:go_default_library",
+        "//pkg/downwardmetrics/scraper:go_default_library",
         "//pkg/healthz:go_default_library",
-        "//pkg/monitoring/domainstats/downwardmetrics:go_default_library",
         "//pkg/monitoring/metrics/common/client:go_default_library",
         "//pkg/monitoring/metrics/virt-handler:go_default_library",
         "//pkg/monitoring/metrics/virt-handler/handler:go_default_library",
