+++ pkg/downwardmetrics/virtio-serial/BUILD.bazel
@@ -6,9 +6,9 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/downwardmetrics/virtio-serial",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/downwardmetrics/scraper:go_default_library",
         "//pkg/downwardmetrics/vhostmd/api:go_default_library",
         "//pkg/ephemeral-disk-utils:go_default_library",
-        "//pkg/monitoring/domainstats/downwardmetrics:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/golang.org/x/time/rate:go_default_library",
     ],
